requirements.txt
python-binance
TA-LIB
numpy
websocket-client
streamlit